requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
python-multipart==0.0.7
httpx==0.26.0
websockets==12.0
pandas==2.2.0
openpyxl==3.1.2
requests==2.31.0
python-dotenv==1.0.1
psutil==5.9.8
GPUtil==1.4.0
openai==1.10.0
